refactor: renamed roles names
