+++ src/main/java/com/produtos/apirest/enums/RoleName.java
@@ -4,6 +4,6 @@ public enum RoleName {
     ROLE_ADMIN,
     ROLE_SECRETARIO,
     //Roles de Teste
-    ROLE_TESTE,
-    ROLE_TESTE2,
+    ROLE_TEST,
+    ROLE_TEST2,
 }
