+++ src/main/java/com/produtos/apirest/service/AreaService.java
@@ -3,7 +3,7 @@ package com.produtos.apirest.service;
 import com.produtos.apirest.models.Area;
 import com.produtos.apirest.models.Expertise;
 import com.produtos.apirest.repository.AreaRepo;
-import com.produtos.apirest.service.excecoes.BusinessRuleException;
+import com.produtos.apirest.service.exceptions.BusinessRuleException;
 import org.hibernate.Hibernate;
 import org.springframework.data.domain.Example;
 import org.springframework.data.domain.ExampleMatcher;
