+++ src/main/java/com/produtos/apirest/service/VeterinarioService.java
@@ -3,7 +3,7 @@ package com.produtos.apirest.service;
 import com.produtos.apirest.models.Expertise;
 import com.produtos.apirest.models.Veterinary;
 import com.produtos.apirest.repository.VeterinaryRepo;
-import com.produtos.apirest.service.excecoes.BusinessRuleException;
+import com.produtos.apirest.service.exceptions.BusinessRuleException;
 import org.springframework.data.domain.Example;
 import org.springframework.data.domain.ExampleMatcher;
 import org.springframework.stereotype.Service;
