+++ src/main/java/com/produtos/apirest/service/exceptions/BusinessRuleException.java
@@ -1,4 +1,4 @@
-package com.produtos.apirest.service.excecoes;
+package com.produtos.apirest.service.exceptions;
 public class BusinessRuleException extends RuntimeException{
     public BusinessRuleException(String msg){
         super(msg);
